# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ALMEAA API
  version: 0.4.0
paths:
...
  /api/v1/questions:
    get:
      operationId: questionBankStaffList
      description: >
        Bounded staff Question Bank list. Default limit 80, maximum 100. Teacher access
        is forcibly scoped to owned/assigned questions. Platform-trainer scope is canonical
        in Content; an exact pathId+subjectId filter may also be authorized by an active
        Organizations teaching assignment. Exact coverage counts are not computed on this normal list path.
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 80 } }
        - { in: query, name: search, schema: { type: string, maxLength: 200 } }
        - { in: query, name: pathId, schema: { type: string } }
        - { in: query, name: subjectId, schema: { type: string } }
        - { in: query, name: mainSkillId, schema: { type: string } }
        - { in: query, name: skillId, schema: { type: string }, description: Repeatable skill filter; any matching skill qualifies. }
        - { in: query, name: linked, schema: { type: boolean } }
        - { in: query, name: difficulty, schema: { type: string } }
        - { in: query, name: type, schema: { type: string, enum: [mcq, true_false, essay] } }
        - { in: query, name: examType, schema: { type: string } }
        - { in: query, name: source, schema: { type: string } }
        - { in: query, name: year, schema: { type: integer } }
        - { in: query, name: workflowStatus, schema: { type: string, enum: [draft, pending_review, approved, rejected, archived] } }
        - { in: query, name: withVideo, schema: { type: boolean } }
        - { in: query, name: withExplanation, schema: { type: boolean } }
      responses:
        "200": { description: Bounded staff summary page with hasMore }
        "400": { description: Invalid filter or pagination }
        "401": { description: Authentication required }
        "403": { description: Staff role required }
    post:
      operationId: questionBankCreate
      description: >
        Create a stable question identity with version 1. Staff-only and CSRF-protected.
        Teachers may author only inside canonical active Content platform-trainer scope or an
        exact active Organizations teaching assignment; admin assignment is rejected outside
        the target teacher's combined authoring scope.
        A question may be text-only, image-only, or text+image. Image bytes are never sent
        through this endpoint; media is referenced by asset ID. AI-readable metadata stays
        server-side in the staff version payload.
      responses:
        "201": { description: Draft question created }
        "400": { description: Invalid question/version payload }
        "403": { description: Staff role or CSRF denied }
        "409": { description: Question code, taxonomy, asset, or hierarchy conflict }
...
  /api/v1/schools:
    get:
      operationId: organizationsListSchools
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { in: query, name: search, schema: { type: string, maxLength: 120 } }
        - { in: query, name: status, schema: { type: string, enum: [active, suspended, archived] } }
      responses:
        "200": { description: Paginated schools visible to the authenticated organization scope }
        "401": { description: Authentication required }
        "403": { description: Organization role or scope denied }
    post:
      operationId: organizationsCreateSchool
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                code: { type: string, maxLength: 64 }
                name: { type: string, minLength: 1, maxLength: 160 }
                metadata: { type: object }
      responses:
        "201": { description: School created }
        "400": { description: Invalid request }
        "401": { description: Authentication required }
        "403": { description: Platform admin role and valid CSRF required }
        "409": { description: School code or normalized school identity conflicts }
...
  /api/v1/courses:
    get:
      operationId: contentListCourses
      description: >
        Authenticated staff list. Default limit 50, maximum 100, returns compact rows plus
        hasMore. Teacher scope is forcibly limited to content they own or are explicitly assigned.
        Platform-trainer scope is canonical in Content; an exact pathId+subjectId filter may also
        be authorized by an active Organizations teaching assignment.
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { in: query, name: pathId, schema: { type: string } }
        - { in: query, name: subjectId, schema: { type: string } }
        - { in: query, name: search, schema: { type: string, maxLength: 160 } }
        - { in: query, name: workflowStatus, schema: { type: string, enum: [draft, pending_review, approved, rejected, archived] } }
      responses:
        "200": { description: Bounded staff course page }
        "400": { description: Invalid filter or pagination }
        "401": { description: Authentication required }
        "403": { description: Staff scope required }
    post:
      operationId: contentCreateCourse
      description: Platform admins and teachers with canonical platform-trainer scope or an active exact school teaching assignment may create courses. Teacher ownership/assignment is server-forced, revenue share remains admin-controlled, and taxonomy/active Media references are server-validated.
      responses:
        "201": { description: Draft course created }
        "400": { description: Invalid course payload }
        "403": { description: Admin/trainer scope or CSRF denied }
        "409": { description: Taxonomy, owner, asset, or relational conflict }
...
  /api/v1/lessons:
    get:
      operationId: contentListLessons
      description: Bounded staff lesson list with owner/assignment authority; platform trainer scope applies globally, while an exact pathId+subjectId filter may be authorized by an active school teaching assignment.
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { in: query, name: pathId, schema: { type: string } }
        - { in: query, name: subjectId, schema: { type: string } }
        - { in: query, name: search, schema: { type: string, maxLength: 160 } }
        - { in: query, name: workflowStatus, schema: { type: string, enum: [draft, pending_review, approved, rejected, archived] } }
      responses:
        "200": { description: Bounded staff lesson page }
        "403": { description: Staff scope required }
    post:
      operationId: contentCreateLesson
      description: Platform admins and teachers with canonical platform-trainer scope or an active exact school teaching assignment may create lessons. Teacher ownership/assignment is server-forced; binary assets remain in Media/R2 and are referenced by ID.
      responses:
        "201": { description: Draft lesson created }
        "400": { description: Invalid lesson payload }
        "403": { description: Admin/trainer scope or CSRF denied }
        "409": { description: Taxonomy, owner, or asset conflict }
...
  /api/v1/library:
    get:
      operationId: contentListLibrary
      description: Bounded staff library list with owner/assignment authority; platform trainer scope applies globally, while an exact pathId+subjectId filter may be authorized by an active school teaching assignment.
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { in: query, name: pathId, schema: { type: string } }
        - { in: query, name: subjectId, schema: { type: string } }
        - { in: query, name: search, schema: { type: string, maxLength: 160 } }
        - { in: query, name: workflowStatus, schema: { type: string, enum: [draft, pending_review, approved, rejected, archived] } }
      responses:
        "200": { description: Bounded staff library page }
        "403": { description: Staff scope required }
    post:
      operationId: contentCreateLibraryItem
      description: Platform admins and teachers with canonical platform-trainer scope or an active exact school teaching assignment may create library items. Teacher ownership/assignment is server-forced and large binaries stay in Media/R2.
      responses:
        "201": { description: Draft library item created }
        "403": { description: Admin/trainer scope or CSRF denied }
        "409": { description: Taxonomy, owner, or asset conflict }
...